Only use filepath.Abs when we're somewhat sure we have a regular path
Otherwise we'll try to run `filepath.Abs` on a url and the results
aren't pretty. Using the code below as an example:

    func main() {
	uri, _ := filepath.Abs("https://example.com/archive.zip")
	fmt.Println(uri)
    }

When running this from the `/tmp` folder, we get:

    $ ./main
    /tmp/https:/example.com/archive.zip

This isn't what we wanted so we get the specific case of something that
looks like an URL and get the absolute path only when needed. This is
also important because `getter.GetAny` will only download something from
the internet when its argument actually starts with `http://` or
`https://`. So with this test we increase the change of whatever reaches
`getter.GetAny` to actually work